+++ src/dz13/Checking_for_a_prime_number.java
@@ -1,4 +1,4 @@
-package src;
+package src.dz13;
 
 import java.util.Scanner;
