+++ src/dz13/Raznost.java
@@ -1,6 +1,8 @@
+package src.dz13;
+
 import java.util.Scanner;
 
-public class Uslovie {
+public class Raznost {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Пожалуйста введите первое число!");
@@ -8,16 +10,11 @@ public class Uslovie {
         System.out.println("Пожалуйста введите второе число!");
         int b = scanner.nextInt();
         int result;
-        if (a > b ) {
-            System.out.println("Данное число больше");
-            result = a;
-        } else if (a < b) {
-            System.out.println("Данное число больше");
-            result = b;
+        if ( a < b ) {
+            result = b - a;
         } else {
-            System.out.println("Оба числа равны");
-            result = 0;
-        }
-        System.out.println(result);
+            result = a - b;
+        } System.out.println("Ответ: " + result);
+
     }
 }
