+++ src/dz13/dz2.java
@@ -1,3 +1,5 @@
+package src.dz13;
+
 public class dz2 {
     
 }
